in progress results of pair programming
