+++ package.json
@@ -65,6 +65,7 @@
     "eslint-plugin-only-warn": "1.0.3",
     "execa": "5.1.1",
     "fast-glob": "3.2.7",
+    "floggy": "^0.3.2",
     "format-imports": "2.4.1",
     "fs-jetpack": "4.2.0",
     "jest": "27.3.1",
